powershell update

+++ redcannary/runtest.ps1
@@ -21,34 +21,35 @@ foreach ($info in $csv)
 	$valid = $info.sigma
 	$name =  $info.name
 	if ($info.os -like '*windows*'){
-		if ($valid -eq 'False') {
-			write-host "Test $name - $technique test : $nmr" 
-			write-host "Make environnement" 
-			Invoke-AtomicTest $technique -TestNumbers $nmr -GetPrereqs
-			$list_channel | Clear-WinEvents -Verbose
-			Start-Sleep -s 5
+        if ($info.executor -ne 'manual'){ 
+		    if ($valid -eq 'False') {
+			    write-host "Test $name - $technique test : $nmr" 
+			    write-host "Disable Realtime Monitoring" 
+			    Set-MpPreference -DisableRealtimeMonitoring 1
+			    write-host "Make environnement" 
+			    Invoke-AtomicTest $technique -TestNumbers $nmr -Cleanup
+			    Invoke-AtomicTest $technique -TestNumbers $nmr -GetPrereqs -TimeoutSeconds 120
+			    $list_channel | Clear-WinEvents -Verbose
+			    Start-Sleep -s 5
 
-			write-host "Start Aurora" 
-			Start-Process C:\aurora-beta\aurora-agent-64.exe -WorkingDirectory C:\aurora-beta\ -ArgumentList "-c agent-config-standard.yml","--minimum-level low","--json","-l c:\Tests\$($technique)_test_$($nmr)_aurora.json"
-			Start-Sleep -s 15
+			    write-host "Start Aurora" 
+			    Start-Process C:\aurora-beta\aurora-agent-64.exe -WorkingDirectory C:\aurora-beta\ -ArgumentList "-c agent-config-standard.yml","--minimum-level low","--json","-l c:\Tests\$($technique)_test_$($nmr)_aurora.json"
+			    Start-Sleep -s 15
 		
-			write-host "Start test" 
-			Invoke-AtomicTest $technique -TestNumbers $nmr
+			    write-host "Start test" 
+			    Invoke-AtomicTest $technique -TestNumbers $nmr -TimeoutSeconds 120
 
-			Start-Sleep -s 5
-			$list_channel | Export-WinEvents -TimeBucket 'Last 5 Minutes' -OutputPath "$($technique)_test_$($nmr)_dataset.json" -Verbose
+			    Start-Sleep -s 5
+			    $list_channel | Export-WinEvents -TimeBucket 'Last 5 Minutes' -OutputPath "$($technique)_test_$($nmr)_dataset.json" -Verbose
 
-			write-host "Stop Aurora" 
-			Stop-Process -name aurora-agent-64
+			    write-host "Stop Aurora" 
+			    Stop-Process -name aurora-agent-64
 		
-			write-host "Cleanup" 
-			Invoke-AtomicTest $technique -TestNumbers $nmr -Cleanup
-			Start-Sleep -s 5
-		} Else {
-			write-host "$name / $technique test: $nmr allready done" -ForegroundColor green
-		}
-	} Else {
-		write-host "$name / $technique test: $nmr  Not windows" -ForegroundColor yellow
-	}
-	
+			    write-host "Cleanup" 
+			    Invoke-AtomicTest $technique -TestNumbers $nmr -Cleanup
+			    Start-Sleep -s 5
+		    } Else { write-host "$name / $technique test: $nmr / OK" -ForegroundColor green }
+		} Else { write-host "$name / $technique test: $nmr / manual test :)" -ForegroundColor DarkRed }
+	} Else { write-host "$name / $technique test: $nmr / not windows :)" -ForegroundColor DarkRed }
 }
+write-host "Good Hunt..." -ForegroundColor green
